backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "prophetai-backend"
version = "0.1.0"
description = "ProphetAI — AI-powered real estate analytics backend"
requires-python = ">=3.11"
dependencies = [
    "django==4.2.9",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.3.1",
    "google-generativeai==0.4.0",
    "pandas==2.1.4",
    "scikit-learn==1.3.2",
    "xgboost==2.0.3",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "Pillow==10.2.0",
    "requests==2.31.0",
    "redis==5.0.1",
    "django-redis==5.4.0",
    "celery==5.3.6",
    "gunicorn==21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-django==4.7.0",
    "factory-boy==3.3.0",
    "black==24.1.1",
    "ruff==0.1.14",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "prophetai.settings"
python_files = ["test_*.py"]
python_classes = ["*Test"]
python_functions = ["test_*"]
